+++ src/IncidentCompass.Application/Intake/IntakeSetup.cs
@@ -19,6 +19,7 @@ internal static class IntakeSetup
         services.TryAddSingleton<SignalNormalizerRegistry>();
 
         services.TryAddScoped<GroundedFactsAssembler>();
+        services.TryAddScoped<OpenFaultNeighborSetRefresher>();
         services.TryAddScoped<FaultGroupingCoordinator>();
 
         services.TryAddScoped<IRequestHandler<IngestSignalCommand, IngestSignalResponse>, IngestSignalCommandHandler>();